+++ class-components/src/page/components/list/list.tsx
@@ -7,8 +7,9 @@ export function List(props: Results): ReactElement {
     <li className={styles.listItem}>
       <p className={styles.liName}>{props.name}</p>
       <p
-        className={styles.description}
-      >{`Pokemon id: ${props.url.slice(34).slice(0, -1)}`}</p>
+        className={
+          styles.description
+        }>{`Pokemon id: ${props.url.slice(34).slice(0, -1)}`}</p>
     </li>
   );
 }
